+++ editor/store/test/effects.js
@@ -1,7 +1,7 @@
 /**
  * External dependencies
  */
-import { noop, set, reduce } from 'lodash';
+import { noop } from 'lodash';
 
 /**
  * WordPress dependencies
@@ -12,7 +12,6 @@ import {
 	registerBlockType,
 	createBlock,
 } from '@wordpress/blocks';
-import apiFetch from '@wordpress/api-fetch';
 
 /**
  * Internal dependencies
@@ -24,15 +23,6 @@ import {
 	selectBlock,
 	removeBlock,
 	createErrorNotice,
-	fetchSharedBlocks,
-	receiveSharedBlocks,
-	receiveBlocks,
-	saveSharedBlock,
-	deleteSharedBlock,
-	removeBlocks,
-	resetBlocks,
-	convertBlockToStatic,
-	convertBlockToShared,
 	setTemplateValidity,
 	editPost,
 } from '../actions';
@@ -557,388 +547,6 @@ describe( 'effects', () => {
 				setTemplateValidity( true ),
 				setupEditorState( post, [], { title: 'A History of Pork' } ),
 			] );
-		} );
-	} );
-
-	describe( 'shared block effects', () => {
-		beforeAll( () => {
-			registerBlockType( 'core/test-block', {
-				title: 'Test block',
-				category: 'common',
-				save: () => null,
-				attributes: {
-					name: { type: 'string' },
-				},
-			} );
-			registerBlockType( 'core/block', {
-				title: 'Shared Block',
-				category: 'common',
-				save: () => null,
-				attributes: {
-					ref: { type: 'string' },
-				},
-			} );
-		} );
-
-		afterAll( () => {
-			unregisterBlockType( 'core/test-block' );
-			unregisterBlockType( 'core/block' );
-		} );
-
-		describe( '.FETCH_SHARED_BLOCKS', () => {
-			const handler = effects.FETCH_SHARED_BLOCKS;
-
-			afterEach( () => {
-				jest.unmock( '@wordpress/api-fetch' );
-			} );
-
-			it( 'should fetch multiple shared blocks', () => {
-				const promise = Promise.resolve( [
-					{
-						id: 123,
-						title: 'My cool block',
-						content: '<!-- wp:test-block {"name":"Big Bird"} /-->',
-					},
-				] );
-				apiFetch.mockReturnValue = promise;
-				set( global, [ 'wp', 'api', 'getPostTypeRoute' ], () => 'blocks' );
-
-				const dispatch = jest.fn();
-				const store = { getState: noop, dispatch };
-
-				handler( fetchSharedBlocks(), store );
-
-				return promise.then( () => {
-					expect( dispatch ).toHaveBeenCalledWith(
-						receiveSharedBlocks( [
-							{
-								sharedBlock: {
-									id: 123,
-									title: 'My cool block',
-									content: '<!-- wp:test-block {"name":"Big Bird"} /-->',
-								},
-								parsedBlock: expect.objectContaining( {
-									name: 'core/test-block',
-									attributes: { name: 'Big Bird' },
-								} ),
-							},
-						] )
-					);
-					expect( dispatch ).toHaveBeenCalledWith( {
-						type: 'FETCH_SHARED_BLOCKS_SUCCESS',
-						id: undefined,
-					} );
-				} );
-			} );
-
-			it( 'should fetch a single shared block', () => {
-				const promise = Promise.resolve( {
-					id: 123,
-					title: 'My cool block',
-					content: '<!-- wp:test-block {"name":"Big Bird"} /-->',
-				} );
-				apiFetch.mockReturnValue = promise;
-				set( global, [ 'wp', 'api', 'getPostTypeRoute' ], () => 'blocks' );
-
-				const dispatch = jest.fn();
-				const store = { getState: noop, dispatch };
-
-				handler( fetchSharedBlocks( 123 ), store );
-
-				return promise.then( () => {
-					expect( dispatch ).toHaveBeenCalledWith(
-						receiveSharedBlocks( [
-							{
-								sharedBlock: {
-									id: 123,
-									title: 'My cool block',
-									content: '<!-- wp:test-block {"name":"Big Bird"} /-->',
-								},
-								parsedBlock: expect.objectContaining( {
-									name: 'core/test-block',
-									attributes: { name: 'Big Bird' },
-								} ),
-							},
-						] )
-					);
-					expect( dispatch ).toHaveBeenCalledWith( {
-						type: 'FETCH_SHARED_BLOCKS_SUCCESS',
-						id: 123,
-					} );
-				} );
-			} );
-
-			it( 'should handle an API error', () => {
-				const promise = Promise.reject( {
-					code: 'unknown_error',
-					message: 'An unknown error occurred.',
-				} );
-				apiFetch.mockReturnValue = promise;
-				set( global, [ 'wp', 'api', 'getPostTypeRoute' ], () => 'blocks' );
-
-				const dispatch = jest.fn();
-				const store = { getState: noop, dispatch };
-
-				handler( fetchSharedBlocks(), store );
-
-				return promise.catch( () => {
-					process.nextTick( () => {
-						expect( dispatch ).toHaveBeenCalledWith( {
-							type: 'FETCH_SHARED_BLOCKS_FAILURE',
-							error: {
-								code: 'unknown_error',
-								message: 'An unknown error occurred.',
-							},
-						} );
-					} );
-				} );
-			} );
-		} );
-
-		describe( '.RECEIVE_SHARED_BLOCKS', () => {
-			const handler = effects.RECEIVE_SHARED_BLOCKS;
-
-			it( 'should receive parsed blocks', () => {
-				const action = receiveSharedBlocks( [
-					{
-						parsedBlock: { clientId: 'broccoli' },
-					},
-				] );
-
-				expect( handler( action ) ).toEqual( receiveBlocks( [
-					{ clientId: 'broccoli' },
-				] ) );
-			} );
-		} );
-
-		describe( '.SAVE_SHARED_BLOCK', () => {
-			const handler = effects.SAVE_SHARED_BLOCK;
-
-			it( 'should save a shared block and swap its id', () => {
-				const promise = Promise.resolve( { id: 456 } );
-				apiFetch.mockReturnValue = promise;
-
-				set( global, [ 'wp', 'api', 'getPostTypeRoute' ], () => 'blocks' );
-
-				const sharedBlock = { id: 123, title: 'My cool block' };
-				const parsedBlock = createBlock( 'core/test-block', { name: 'Big Bird' } );
-
-				const state = reduce( [
-					receiveSharedBlocks( [ { sharedBlock, parsedBlock } ] ),
-					receiveBlocks( [ parsedBlock ] ),
-				], reducer, undefined );
-
-				const dispatch = jest.fn();
-				const store = { getState: () => state, dispatch };
-
-				handler( saveSharedBlock( 123 ), store );
-
-				return promise.then( () => {
-					expect( dispatch ).toHaveBeenCalledWith( {
-						type: 'SAVE_SHARED_BLOCK_SUCCESS',
-						id: 123,
-						updatedId: 456,
-					} );
-				} );
-			} );
-
-			it( 'should handle an API error', ( done ) => {
-				const promise = Promise.reject( {} );
-				apiFetch.mockReturnValue = promise;
-				set( global, [ 'wp', 'api', 'getPostTypeRoute' ], () => 'blocks' );
-
-				const sharedBlock = { id: 123, title: 'My cool block' };
-				const parsedBlock = createBlock( 'core/test-block', { name: 'Big Bird' } );
-
-				const state = reduce( [
-					receiveSharedBlocks( [ { sharedBlock, parsedBlock } ] ),
-					receiveBlocks( [ parsedBlock ] ),
-				], reducer, undefined );
-
-				const dispatch = jest.fn();
-				const store = { getState: () => state, dispatch };
-
-				handler( saveSharedBlock( 123 ), store );
-
-				promise.catch( () => {
-					process.nextTick( () => {
-						expect( dispatch ).toHaveBeenCalledWith( {
-							type: 'SAVE_SHARED_BLOCK_FAILURE',
-							id: 123,
-						} );
-						done();
-					} );
-				} );
-			} );
-		} );
-
-		describe( '.DELETE_SHARED_BLOCK', () => {
-			const handler = effects.DELETE_SHARED_BLOCK;
-
-			it( 'should delete a shared block', () => {
-				const promise = Promise.resolve( {} );
-				apiFetch.mockReturnValue = promise;
-				set( global, [ 'wp', 'api', 'getPostTypeRoute' ], () => 'blocks' );
-
-				const associatedBlock = createBlock( 'core/block', { ref: 123 } );
-				const sharedBlock = { id: 123, title: 'My cool block' };
-				const parsedBlock = createBlock( 'core/test-block', { name: 'Big Bird' } );
-
-				const state = reduce( [
-					resetBlocks( [ associatedBlock ] ),
-					receiveSharedBlocks( [ { sharedBlock, parsedBlock } ] ),
-					receiveBlocks( [ parsedBlock ] ),
-				], reducer, undefined );
-
-				const dispatch = jest.fn();
-				const store = { getState: () => state, dispatch };
-
-				handler( deleteSharedBlock( 123 ), store );
-
-				expect( dispatch ).toHaveBeenCalledWith( {
-					type: 'REMOVE_SHARED_BLOCK',
-					id: 123,
-					optimist: expect.any( Object ),
-				} );
-
-				expect( dispatch ).toHaveBeenCalledWith(
-					removeBlocks( [ associatedBlock.clientId, parsedBlock.clientId ] )
-				);
-
-				return promise.then( () => {
-					expect( dispatch ).toHaveBeenCalledWith( {
-						type: 'DELETE_SHARED_BLOCK_SUCCESS',
-						id: 123,
-						optimist: expect.any( Object ),
-					} );
-				} );
-			} );
-
-			it( 'should handle an API error', ( done ) => {
-				const promise = Promise.reject( {} );
-				apiFetch.mockReturnValue = promise;
-				set( global, [ 'wp', 'api', 'getPostTypeRoute' ], () => 'blocks' );
-
-				const sharedBlock = { id: 123, title: 'My cool block' };
-				const parsedBlock = createBlock( 'core/test-block', { name: 'Big Bird' } );
-
-				const state = reduce( [
-					receiveSharedBlocks( [ { sharedBlock, parsedBlock } ] ),
-					receiveBlocks( [ parsedBlock ] ),
-				], reducer, undefined );
-
-				const dispatch = jest.fn();
-				const store = { getState: () => state, dispatch };
-
-				handler( deleteSharedBlock( 123 ), store );
-
-				promise
-					.catch( () => {
-						process.nextTick( () => {
-							expect( dispatch ).toHaveBeenCalledWith( {
-								type: 'DELETE_SHARED_BLOCK_FAILURE',
-								id: 123,
-								optimist: expect.any( Object ),
-							} );
-							done();
-						} );
-					} );
-			} );
-
-			it( 'should not save shared blocks with temporary IDs', () => {
-				const sharedBlock = { id: 'shared1', title: 'My cool block' };
-				const parsedBlock = createBlock( 'core/test-block', { name: 'Big Bird' } );
-
-				const state = reduce( [
-					receiveSharedBlocks( [ { sharedBlock, parsedBlock } ] ),
-					receiveBlocks( [ parsedBlock ] ),
-				], reducer, undefined );
-
-				const dispatch = jest.fn();
-				const store = { getState: () => state, dispatch };
-
-				handler( deleteSharedBlock( 'shared1' ), store );
-
-				expect( dispatch ).not.toHaveBeenCalled();
-			} );
-		} );
-
-		describe( '.CONVERT_BLOCK_TO_STATIC', () => {
-			const handler = effects.CONVERT_BLOCK_TO_STATIC;
-
-			it( 'should convert a shared block into a static block', () => {
-				const associatedBlock = createBlock( 'core/block', { ref: 123 } );
-				const sharedBlock = { id: 123, title: 'My cool block' };
-				const parsedBlock = createBlock( 'core/test-block', { name: 'Big Bird' } );
-
-				const state = reduce( [
-					resetBlocks( [ associatedBlock ] ),
-					receiveSharedBlocks( [ { sharedBlock, parsedBlock } ] ),
-					receiveBlocks( [ parsedBlock ] ),
-				], reducer, undefined );
-
-				const dispatch = jest.fn();
-				const store = { getState: () => state, dispatch };
-
-				handler( convertBlockToStatic( associatedBlock.clientId ), store );
-
-				expect( dispatch ).toHaveBeenCalledWith( {
-					type: 'REPLACE_BLOCKS',
-					clientIds: [ associatedBlock.clientId ],
-					blocks: [
-						expect.objectContaining( {
-							name: 'core/test-block',
-							attributes: { name: 'Big Bird' },
-						} ),
-					],
-					time: expect.any( Number ),
-				} );
-			} );
-		} );
-
-		describe( '.CONVERT_BLOCK_TO_SHARED', () => {
-			const handler = effects.CONVERT_BLOCK_TO_SHARED;
-
-			it( 'should convert a static block into a shared block', () => {
-				const staticBlock = createBlock( 'core/block', { ref: 123 } );
-				const state = reducer( undefined, resetBlocks( [ staticBlock ] ) );
-
-				const dispatch = jest.fn();
-				const store = { getState: () => state, dispatch };
-
-				handler( convertBlockToShared( staticBlock.clientId ), store );
-
-				expect( dispatch ).toHaveBeenCalledWith(
-					receiveSharedBlocks( [ {
-						sharedBlock: {
-							id: expect.stringMatching( /^shared/ ),
-							clientId: staticBlock.clientId,
-							title: 'Untitled shared block',
-						},
-						parsedBlock: staticBlock,
-					} ] )
-				);
-
-				expect( dispatch ).toHaveBeenCalledWith(
-					saveSharedBlock( expect.stringMatching( /^shared/ ) ),
-				);
-
-				expect( dispatch ).toHaveBeenCalledWith( {
-					type: 'REPLACE_BLOCKS',
-					clientIds: [ staticBlock.clientId ],
-					blocks: [
-						expect.objectContaining( {
-							name: 'core/block',
-							attributes: { ref: expect.stringMatching( /^shared/ ) },
-						} ),
-					],
-					time: expect.any( Number ),
-				} );
-
-				expect( dispatch ).toHaveBeenCalledWith(
-					receiveBlocks( [ staticBlock ] ),
-				);
-			} );
 		} );
 	} );
 } );
